Cleanup old dags and old tests
