+++ dags/sources/config/shared_constants.py
@@ -1,9 +1,3 @@
-# DEPRECATED DagRun statuts
-DAGRUN_TOVALIDATE = "TO_VALIDATE"
-DAGRUN_TOINSERT = "TO_INSERT"
-DAGRUN_REJECTED = "REJECTED"
-DAGRUN_FINISHED = "FINISHED"
-
 # Suggestion statut
 SUGGESTION_AVALIDER = "AVALIDER"
 SUGGESTION_REJETEE = "REJETEE"
